test: split sampler suite and remove conftest path hack
